+++ src/components/client/index.vue
@@ -215,11 +215,11 @@ th.sortable {
 
 <script type="text/javascript">
 import { reactive, toRefs, computed, watch } from 'vue';
-import useFilters from '@/components/mixins/filters';
-import useDefaultOptions from '@/components/mixins/default-options';
-import { useToggle, useComputedColumns } from '@/components/mixins/methods';
+import useFormatters from '@/components/common/formatters';
+import useDefaultOptions from '@/components/common/default-options';
+import { useToggle, useComputedColumns } from '@/components/common/methods';
 import Pagination from '@/components/Pagination.vue';
-import ActionsCell from '@/components/mixins/ActionsCell.vue';
+import ActionsCell from '@/components/common/ActionsCell.vue';
 import { usePagination, useSort, useSelect, useFilter, useGroups } from './methods';
 
 export default {
@@ -366,7 +366,7 @@ export default {
 
     return {
       ...toRefs(state),
-      ...useFilters(),
+      ...useFormatters(),
       ...useToggle(state, context),
       ...useComputedColumns({ columns: props.columns, opts, data: props.data }),
       ...usePagination(context, state, filteredData),
