+++ src/components/common/methods.js
